requirements.txt
Pillow
Django
Djangorestframework
Django-ckeditor
Django-mptt